+++ packages/cojson/CHANGELOG.md
@@ -1,5 +1,11 @@
 # cojson
 
+## 0.6.0
+
+### Minor Changes
+
+- Make addMember and removeMember take loaded Accounts instead of just IDs
+
 ## 0.5.2
 
 ### Patch Changes