+++ packages/jazz-autosub/CHANGELOG.md
@@ -1,5 +1,16 @@
 # jazz-autosub
 
+## 0.6.0
+
+### Minor Changes
+
+- Make addMember and removeMember take loaded Accounts instead of just IDs
+
+### Patch Changes
+
+- Updated dependencies
+  - cojson@0.6.0
+
 ## 0.5.0
 
 ### Minor Changes