+++ packages/jazz-browser-media-images/CHANGELOG.md
@@ -1,5 +1,14 @@
 # jazz-browser-media-images
 
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies
+  - jazz-autosub@0.6.0
+  - jazz-browser@0.6.0
+  - cojson@0.6.0
+
 ## 0.5.0
 
 ### Minor Changes