+++ packages/jazz-react-auth-local/CHANGELOG.md
@@ -1,5 +1,13 @@
 # jazz-react-auth-local
 
+## 0.4.17
+
+### Patch Changes
+
+- Updated dependencies
+  - jazz-browser-auth-local@0.5.0
+  - jazz-react@0.5.1
+
 ## 0.4.16
 
 ### Patch Changes